restaurant_search_engine_test/requirements.txt
httpx>=0.27
google-generativeai>=0.8
python-dotenv>=1.0
rich>=13.0
